create base smart contract

+++ web3.0/smart_contract/contracts/Transactions.sol
@@ -4,12 +4,19 @@ pragma solidity ^0.8.0;
 contract Transactions {
     uint256 transactionCount;
 
-    event Transfer(address _from, address _to, uint256 _amount, string _message, uint256 _timestamp, string _keyword);
+    event Transfer(
+        address _from,
+        address _to,
+        uint256 _amount,
+        string _message,
+        uint256 _timestamp,
+        string _keyword
+    );
 
     struct TransferStruct {
         address sender;
         address receiver;
-        uint amount;
+        uint256 amount;
         string message;
         uint256 timestamp;
         string keyword;
@@ -17,13 +24,38 @@ contract Transactions {
 
     TransferStruct[] transactions;
 
-    function addToBlockchain(address payable _receiver, uint _amount, string memory _message, string memory _keyword) public {
+    function addToBlockchain(
+        address payable _receiver,
+        uint256 _amount,
+        string memory _message,
+        string memory _keyword
+    ) public {
         transactionCount += 1;
-        transactions.push(TransferStruct(msg.sender, _receiver, _amount, _message, block.timestamp, _keyword));
-        emit Transfer(msg.sender, _receiver, _amount, _message, block.timestamp, _keyword);
+        transactions.push(
+            TransferStruct(
+                msg.sender,
+                _receiver,
+                _amount,
+                _message,
+                block.timestamp,
+                _keyword
+            )
+        );
+        emit Transfer(
+            msg.sender,
+            _receiver,
+            _amount,
+            _message,
+            block.timestamp,
+            _keyword
+        );
     }
 
-    function getAllTransactions() public view returns (TransferStruct[] memory) {
+    function getAllTransactions()
+        public
+        view
+        returns (TransferStruct[] memory)
+    {
         return transactions;
     }
